requirements.txt
Django
django-ninja
django-cors-headers
